refacto of previous PR
