--- src/PHPStan/CustomRules/CustomRuleErrorMessage.php
@@ -1,40 +1,0 @@
-<?php
-
-/*
- * Copyright 2005 - 2022 Centreon (https://www.centreon.com/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * For more information : contact@centreon.com
- */
-
-declare(strict_types=1);
-
-namespace Centreon\PHPStan\CustomRules;
-
-/**
- * This class defines static method to build custom error message in PHPStan
- */
-class CustomRuleErrorMessage
-{
-    /**
-     * This method constructs the error message string displayed by PHPStan.
-     *
-     * @param string $varName
-     * @return string
-     */
-    public static function buildErrorMessage(string $errMessage, string $varName = ''): string
-    {
-        return empty($varName) ? '[CENTREON-RULE]: ' . $errMessage : '[CENTREON-RULE]: ' . $varName . ' ' . $errMessage;
-    }
-}
